Move objects service to dedicated module
